LayoutTests/http/tests/security/cross-origin-css-1.html

<!DOCTYPE html>
<html>
<head>
<title>Cross-origin CSS</title>

<script src="/w3c/resources/testharness.js"></script>
<script src="/w3c/resources/testharnessreport.js"></script>

<link rel="stylesheet" href="/resources/redirect.php?url=http://localhost:8000/security/resources/xorigincss1.html"></link>

<script>
function getBackgroundColorForId(id) {
    return window.getComputedStyle(document.getElementById(id), null).getPropertyValue('background-color')
}

var onloadTest = async_test("Testing cross-origin and MIME behavior for CSS.");

window.onload = function () {
    test(function () {
        assert_equals(getBackgroundColorForId('id1'), 'rgba(0, 0, 0, 0)');
    }, 'xorigincss1.html should not be loaded via &lt;link&gt;.');

    onloadTest.done();
};
</script>
</head>
<body>
    <div id="id1" class="id1"></div>
</body>
</html>
